server/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
chromadb==0.4.22
google-auth==2.26.1
python-dotenv==1.0.0
openai==1.7.0
tiktoken==0.5.2
scikit-learn==1.3.2
psycopg2==2.9.9
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
llama-index==0.9.45.post1
docx2txt==0.8
pypdf==4.0.1
pgvector==0.2.5
psycopg2-binary==2.9.9
asyncpg==0.29.0